+++ src/react/hooks/__tests__/useQuery.test.tsx
@@ -9224,8 +9224,8 @@ describe("useQuery Hook", () => {
           },
         },
         dataState: "streaming",
-        loading: false,
-        networkStatus: NetworkStatus.ready,
+        loading: true,
+        networkStatus: NetworkStatus.streaming,
         previousData: undefined,
         variables: {},
       });
@@ -9338,8 +9338,8 @@ describe("useQuery Hook", () => {
           ],
         },
         dataState: "streaming",
-        loading: false,
-        networkStatus: NetworkStatus.ready,
+        loading: true,
+        networkStatus: NetworkStatus.streaming,
         previousData: undefined,
         variables: {},
       });
@@ -9376,8 +9376,8 @@ describe("useQuery Hook", () => {
           ],
         },
         dataState: "streaming",
-        loading: false,
-        networkStatus: NetworkStatus.ready,
+        loading: true,
+        networkStatus: NetworkStatus.streaming,
         previousData: {
           greetings: [
             { message: "Hello world", __typename: "Greeting" },
@@ -9534,8 +9534,8 @@ describe("useQuery Hook", () => {
           ],
         },
         dataState: "streaming",
-        loading: false,
-        networkStatus: NetworkStatus.ready,
+        loading: true,
+        networkStatus: NetworkStatus.streaming,
         previousData: undefined,
         variables: {},
       });
@@ -9592,8 +9592,8 @@ describe("useQuery Hook", () => {
           ],
         },
         dataState: "streaming",
-        loading: false,
-        networkStatus: NetworkStatus.ready,
+        loading: true,
+        networkStatus: NetworkStatus.streaming,
         previousData: {
           allProducts: [
             {
@@ -9680,8 +9680,8 @@ describe("useQuery Hook", () => {
           },
         },
         dataState: "streaming",
-        loading: false,
-        networkStatus: NetworkStatus.ready,
+        loading: true,
+        networkStatus: NetworkStatus.streaming,
         previousData: undefined,
         variables: {},
       });
@@ -9814,8 +9814,8 @@ describe("useQuery Hook", () => {
           },
         },
         dataState: "streaming",
-        loading: false,
-        networkStatus: NetworkStatus.ready,
+        loading: true,
+        networkStatus: NetworkStatus.streaming,
         previousData: undefined,
         variables: {},
       });
@@ -9965,8 +9965,8 @@ describe("useQuery Hook", () => {
           },
         },
         dataState: "streaming",
-        loading: false,
-        networkStatus: NetworkStatus.ready,
+        loading: true,
+        networkStatus: NetworkStatus.streaming,
         previousData: undefined,
         variables: {},
       });
@@ -10143,8 +10143,8 @@ describe("useQuery Hook", () => {
           },
         },
         dataState: "streaming",
-        loading: false,
-        networkStatus: NetworkStatus.ready,
+        loading: true,
+        networkStatus: NetworkStatus.streaming,
         previousData: {
           greeting: {
             __typename: "Greeting",
@@ -10273,8 +10273,8 @@ describe("useQuery Hook", () => {
           },
         },
         dataState: "streaming",
-        loading: false,
-        networkStatus: NetworkStatus.ready,
+        loading: true,
+        networkStatus: NetworkStatus.streaming,
         previousData: {
           greeting: {
             __typename: "Greeting",